Split docoding to doencode / dodecode per coding standard
