+++ src/agora/agora.h
@@ -19,8 +19,9 @@
 #include "concurrent_queue_wrapper.h"
 #include "concurrentqueue.h"
 #include "config.h"
-#include "docoding.h"
+#include "dodecode.h"
 #include "dodemul.h"
+#include "doencode.h"
 #include "dofft.h"
 #include "doprecode.h"
 #include "dozf.h"
