--- src/agora/docoding.h
@@ -1,71 +1,0 @@
-/**
- * @file docoding.h
- * @brief Declaration file for the Docoding class.  Includes the DoEncode and
- * DoDecode classes.
- */
-
-#ifndef DOCODING_H_
-#define DOCODING_H_
-
-#include <armadillo>
-#include <cstdio>
-#include <cstring>
-#include <iostream>
-#include <vector>
-
-#include "buffer.h"
-#include "concurrentqueue.h"
-#include "config.h"
-#include "doer.h"
-#include "encoder.h"
-#include "gettime.h"
-#include "iobuffer.h"
-#include "memory_manage.h"
-#include "modulation.h"
-#include "phy_stats.h"
-#include "scrambler.h"
-#include "stats.h"
-#include "symbols.h"
-#include "utils_ldpc.h"
-
-class DoEncode : public Doer {
- public:
-  DoEncode(Config* in_config, int in_tid, Table<int8_t>& in_raw_data_buffer,
-           Table<int8_t>& in_encoded_buffer, Stats* in_stats_manager);
-  ~DoEncode() override;
-
-  EventData Launch(size_t tag) override;
-
- private:
-  Table<int8_t>& raw_data_buffer_;
-  int8_t* parity_buffer_;  // Intermediate buffer to hold LDPC encoding parity
-
-  // Intermediate buffer to hold LDPC encoding output
-  int8_t* encoded_buffer_temp_;
-  Table<int8_t>& encoded_buffer_;
-  DurationStat* duration_stat_;
-  int8_t* scrambler_buffer_;
-
-  std::unique_ptr<AgoraScrambler::Scrambler> scrambler_;
-};
-
-class DoDecode : public Doer {
- public:
-  DoDecode(Config* in_config, int in_tid,
-           PtrCube<kFrameWnd, kMaxSymbols, kMaxUEs, int8_t>& demod_buffers,
-           PtrCube<kFrameWnd, kMaxSymbols, kMaxUEs, uint8_t>& decoded_buffers,
-           PhyStats* in_phy_stats, Stats* in_stats_manager);
-  ~DoDecode() override;
-
-  EventData Launch(size_t tag) override;
-
- private:
-  int16_t* resp_var_nodes_;
-  PtrCube<kFrameWnd, kMaxSymbols, kMaxUEs, int8_t>& demod_buffers_;
-  PtrCube<kFrameWnd, kMaxSymbols, kMaxUEs, uint8_t>& decoded_buffers_;
-  PhyStats* phy_stats_;
-  DurationStat* duration_stat_;
-  std::unique_ptr<AgoraScrambler::Scrambler> scrambler_;
-};
-
-#endif  // DOCODING_H_
